Update to L18 & Complate

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
 import { MessageService } from '../_services/message.service';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -11,12 +12,12 @@ import { MessageService } from '../_services/message.service';
 export class MessagesComponent implements OnInit {
   messages: Message[] = [];
   pagination: Pagination;
-  container: 'Unread';
+  container: string = "Unread";
   pageNumber = 1;
   pageSize = 5;
   loadingFlag = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService : ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -26,15 +27,20 @@ export class MessagesComponent implements OnInit {
     this.loadingFlag = false;
     this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(response => {
       this.messages = response.result;
+      //console.log(response.pagination.totalItems);
       this.pagination = response.pagination;
       this.loadingFlag = false;
     })
   }
   
   deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(() =>{
-      this.messages.splice(this.messages.findIndex(m => m.id  === id), 1);
-    })
+    this.confirmService.confirm('Confirm Delete Message', 'This can not be undone').subscribe(result => {
+      if(result){
+        this.messageService.deleteMessage(id).subscribe(() =>{
+          this.messages.splice(this.messages.findIndex(m => m.id  === id), 1);
+        })
+      }
+    });
   }
